ERC7984Freezable extension behave like ERC7984
